Rename `lexer_peek_helpers`
