+++ src/lexer_peek_helpers.c
@@ -1,9 +1,13 @@
-#include "include/lexer_peek.h"
+#include "include/lexer_peek_helpers.h"
 #include "include/lexer_struct.h"
 #include "include/macros.h"
 
 #include <ctype.h>
 #include <stdbool.h>
+
+char lexer_backtrack(lexer_T* lexer, int offset) {
+  return lexer->source[MAX(lexer->current_position - offset, 0)];
+}
 
 char lexer_peek(lexer_T* lexer, int offset) {
   return lexer->source[MIN(lexer->current_position + offset, lexer->source_length)];
@@ -44,8 +48,4 @@ bool lexer_peek_erb_percent_close_tag(lexer_T* lexer, int offset) {
 bool lexer_peek_erb_end(lexer_T* lexer, int offset) {
   return (lexer_peek_erb_close_tag(lexer, offset) || lexer_peek_erb_dash_close_tag(lexer, offset) ||
           lexer_peek_erb_percent_close_tag(lexer, offset));
-}
-
-char lexer_backtrack(lexer_T* lexer, int offset) {
-  return lexer->source[MAX(lexer->current_position - offset, 0)];
 }
